% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R, R/cAMARETTO_HTMLreport.R
\name{create_hgt_datatable}
\alias{create_hgt_datatable}
\title{Title create_hgt_datatable}
\usage{
create_hgt_datatable(output_hgt, com_table = FALSE, ComNr = 1)

create_hgt_datatable(output_hgt, com_table = FALSE, ComNr = 1)
}
\arguments{
\item{output_hgt}{GSEA test dataframe from HyperGeoEnrichmentTest function.}

\item{com_table}{}

\item{ComNr}{}

\item{module_table}{If TRUE, makes the ModuleNr datatable, If FALSE, makes the index page datatable.}

\item{ModuleNr}{The module number.}

\item{output_hgt}{}
}
\value{
result

DataTable
}
\description{
Title create_hgt_datatable

Title create_hgt_datatable
}
