% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{create_hgt_datatableC}
\alias{create_hgt_datatableC}
\title{Title create_hgt_datatableC}
\usage{
create_hgt_datatableC(output_hgt, com_table = FALSE, ComNr = 1)
}
\arguments{
\item{output_hgt}{}

\item{com_table}{}

\item{ComNr}{}
}
\value{
DataTable
}
\description{
Title create_hgt_datatableC
}
