% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R, R/cAMARETTO_Results.R
\name{GmtFromModules}
\alias{GmtFromModules}
\title{GmtFromModules}
\usage{
GmtFromModules(AMARETTOresults, gmt_file, run, Drivers = FALSE)

GmtFromModules(AMARETTOresults, gmt_file, run, Drivers = FALSE)
}
\arguments{
\item{AMARETTOresults}{List output from AMARETTO_Run().}

\item{gmt_file}{A gmtfilename}

\item{run}{A runname}

\item{Drivers}{Add drivers to the gmt-file}

\item{driverGSEA}{if TRUE , module driver genes will also be added to module target genes for GSEA.}

\item{AMARETTOresults}{A AMARETTO_Run output.}
}
\value{
result

Creates a gmt file for a AMARETTO run
}
\description{
GmtFromModules

GmtFromModules
}
\keyword{internal}
