% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R, R/cAMARETTO_Results.R
\name{HyperGTestGeneEnrichment}
\alias{HyperGTestGeneEnrichment}
\title{Hyper Geometric Geneset Enrichement Test}
\usage{
HyperGTestGeneEnrichment(gmtfile1, gmtfile2, runname1, runname2, NrCores,
  ref.numb.genes = 45956)

HyperGTestGeneEnrichment(gmtfile1, gmtfile2, runname1, runname2, NrCores,
  ref.numb.genes = 45956)
}
\arguments{
\item{gmtfile1}{A gmtfilename that you want to compare}

\item{gmtfile2}{A second gmtfile to compare with.}

\item{runname1}{name of the first dataset.}

\item{runname2}{name of the second dataset.}

\item{NrCores}{Number of cores used for parallelization.}

\item{ref.numb.genes}{The total number of genes teste, standard equal to 45 956 (MSIGDB standard).}

\item{gmtfile}{The gmt file with reference gene set.}

\item{testgmtfile}{The gmt file with gene sets to test. In our case, the gmt file of the modules.}

\item{NrCores}{Number of cores for parallel computing.}

\item{ref.numb.genes}{The reference number of genes.}
}
\value{
result

Creates resultfile with p-values and padj when comparing two gmt files with a hyper geometric test.
}
\description{
Calculates the p-values for unranked gene set enrichment based on two gmt files as input and the hyper geometric test.
}
\keyword{internal}
