% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaretto_htmlreport.R, R/cAMARETTO_Results.R
\name{readGMT}
\alias{readGMT}
\title{readGMT}
\usage{
readGMT(filename)

readGMT(filename)
}
\arguments{
\item{filename}{}

\item{filename}{A gmtfilename}
}
\value{
result

Reads a gmt file
}
\description{
readGMT

readGMT
}
\examples{
readGMT("file.gmt")
}
\keyword{internal}
